requirements.txt
python-telegram-bot==21.3
anthropic>=0.25.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
